Make client poll server for finished programming phase

+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/NetworkController.java
@@ -11,7 +11,7 @@ public class NetworkController {
 	}
 
 	public void sendData(Player player){
-		MovePlayedRest.requestNewMove(gameController.board.lobby.getRounds().intValue(), player.getProgramField(0).getCardName(),
+		MovePlayedRest.requestNewMove(gameController.board.lobby.getRounds(), player.getProgramField(0).getCardName(),
 			player.getProgramField(1).getCardName(), player.getProgramField(2).getCardName(), player.getProgramField(3).getCardName(),
 			player.getProgramField(4).getCardName(), gameController.board.lobby.getId(), player.getNetworkId());
 	}
