+++ src/main/java/dk/dtu/compute/se/pisd/roborally/net/MovePlayedRest.java
@@ -15,7 +15,7 @@ import static dk.dtu.compute.se.pisd.roborallyserver.controller.PlayerController
 public class MovePlayedRest {
 
 	public static final RestTemplate restTemplate = new RestTemplate();
-	public static MovesPlayed requestNewMove(int round, String move1, String move2, String move3, String move4, String move5, Long lobby_id, Long player_id) {
+	public static MovesPlayed requestNewMove(Long round, String move1, String move2, String move3, String move4, String move5, Long lobby_id, Long player_id) {
 		HttpEntity<MovesPlayedController.NewMovesPlayBody> request = new HttpEntity<>(new NewMovesPlayBody(round, move1, move2, move3, move4, move5, lobby_id, player_id));
 		ResponseEntity<MovesPlayed> response = restTemplate
 			.exchange(SERVER_HTTPURL + "movesplayed/newmovesplayed", HttpMethod.POST, request, MovesPlayed.class);
